+++ tests/test_post_form.py
@@ -1,5 +1,7 @@
 import pytest
 
+from django.utils import timezone
+
 from cast.forms import PostForm
 
 
@@ -47,3 +49,32 @@ class TestPostForm:
         form.cleaned_data = cleaned_data
         form._clean_chaptermarks(cleaned_data)
         assert "chaptermarks" in form.errors
+
+    @pytest.mark.django_db
+    def test_post_form_clean_chaptermarks_empty(self, post, audio):
+        # prepare post
+        post.podcast_audio = audio
+
+        # create text for chaptermarks area
+        chaptermarks_text = ""
+
+        # create form + test
+        form = PostForm(instance=post)
+        cleaned_data = {"chaptermarks": chaptermarks_text}
+        form.cleaned_data = cleaned_data
+        form._clean_chaptermarks(cleaned_data)
+        assert len(form.errors) == 0
+
+    def test_safari_pub_date(self, post_data):
+        safari_pub_date = "27.05.2019 14:56:58"
+        post_data["pub_date"] = safari_pub_date
+        post_form = PostForm(post_data)
+        assert post_form.is_valid()
+
+    def test_chrome_empty_pub_date_and_pubished(self, post_data):
+        chrome_pub_date = ""
+        post_data["pub_date"] = chrome_pub_date
+        post_data["is_published"] = True
+        post_form = PostForm(post_data)
+        assert post_form.is_valid()
+        assert (timezone.now() - post_form.cleaned_data["pub_date"]).seconds == 0